Update GatewayAPI to v1.4.0 (#4166)
Problem: We don't currently support the latest version of the Gateway API

Solution: Upgrade to v1.4.0, including BackendTLSPolicy promotion -> v1 and other Policy type promotion -> v1. See the full changelog. Also enables the conformance tests for BackendTLSPolicy.

This also requires upgrading the inference extension from 1.0.2 -> 1.1.0 to resolve dependency conflicts.

+++ internal/controller/provisioner/objects.go
@@ -149,7 +149,7 @@ func (p *NginxProvisioner) buildNginxResourceObjects(
 
 	ports := make(map[int32]struct{})
 	for _, listener := range gateway.Spec.Listeners {
-		ports[int32(listener.Port)] = struct{}{}
+		ports[listener.Port] = struct{}{}
 	}
 
 	// Create separate copies of objectMeta for service and deployment to avoid shared map references
